requirements.txt
beautifulsoup4==4.14.2
chromadb==1.3.7
fastapi==0.124.0
langchain==1.0.3
langchain-anthropic==1.0.1
langchain-community==0.4.1
langchain-core==1.0.2
langchain-openai==1.0.1
langchain-text-splitters==1.0.0
langgraph==1.0.2
SQLAlchemy>=2.0
aiomysql
alembic
email-validator
jinja2
pyjwt
pydantic==2.12.3
python-dotenv==1.2.1
readability-lxml==0.8.4.1
requests==2.32.5
tiktoken==0.12.0
trafilatura==2.0.0
uvicorn==0.38.0
